Mandelbro set
Returns the number of iterations needed to determine if z(n + 1) = z(n) * z(n) + c remains bounded where z(0) = 0 + 0i. z(n + 1) is bounded if its magnitude is less than or equal to 2. Returns 1 if the magnitude of c is greater than 2. Returns max if the magnitude of z(n + 1) is less than or equal to 2 after max iterations.